+++ migrations/001_Create Table Household.sql
@@ -1,7 +1,16 @@
 
+/*************
+*   @UP
+*************/
 CREATE TABLE public.households (
   id serial NOT NULL,
   planning_center_id int NOT NULL,
   name varchar(128) NOT NULL,
   CONSTRAINT pk_household PRIMARY KEY (id)
 );
+
+/***************
+*   @DOWN 
+****************/
+
+DROP TABLE public.households;