+++ src/Console/Configuration.fs
@@ -1,29 +1,38 @@
 module Configuration
-
-open Argu
-open dotenv.net
-
-type Arguments =
-  | [<CustomAppSettings "DB_CONNECTION_STRING"; AltCommandLine "-c">] Connection_String of path:string
-  | [<CustomAppSettings "MIGRATIONS_FOLDER"; AltCommandLine "-m";>] Migrations_Folder of path:string
-
-  interface IArgParserTemplate with 
-    member self.Usage =
-      match self with
-      | Connection_String _ -> "The database connection string."
-      | Migrations_Folder _ -> "Specifies the location the migration scripts folder."
-
-let configure argv =
-  DotEnv.Load(DotEnvOptions(probeLevelsToSearch = 5, probeForEnv = false, ignoreExceptions = true, trimValues = true))
-
-  let errorHandler = ProcessExiter(colorizer = function ErrorCode.HelpText -> None | _ -> Some System.ConsoleColor.Red)
-  let parser = ArgumentParser.Create<Arguments>(programName = "console.exe", errorHandler = errorHandler)
+  
+  open Argu
+  open dotenv.net
+    
+  type Arguments =
+    | [<CustomAppSettings "DB_CONNECTION_STRING"; AltCommandLine "-c">] Connection_String of path:string
+    | [<CustomAppSettings "MIGRATIONS_FOLDER"; AltCommandLine "-m";>] Migrations_Folder of path:string
+    | [<CliPrefix(CliPrefix.None); First; Unique>]Up
+    | [<CliPrefix(CliPrefix.None); First; Unique>]Down of count:uint
+    interface IArgParserTemplate with 
+      member self.Usage =
+        match self with
+        | Connection_String _ -> "The database connection string."
+        | Migrations_Folder _ -> "Specifies the location the migration scripts folder."
+        | Up _ -> "(default) Run all outstanding UP migrations"
+        | Down _ -> "Revert <count> of run migrations"
 
 
-  let options = parser.Parse (inputs = argv, configurationReader = EnvironmentVariableConfigurationReader ())
+  let configure argv =
+    DotEnv.Load(DotEnvOptions(probeLevelsToSearch = 5, probeForEnv = false, ignoreExceptions = true, trimValues = true))
 
-  // Return a record with values or default's set
-  {|
-    ConnectionString = options.GetResult Connection_String
-    MigrationsFolder = options.GetResult (Migrations_Folder, defaultValue = @"..\..\..\..\..\migrations")
-  |}
+    let errorHandler = ProcessExiter(colorizer = function ErrorCode.HelpText -> None | _ -> Some System.ConsoleColor.Red)
+    let parser = ArgumentParser.Create<Arguments>(programName = "console.exe", errorHandler = errorHandler)
+
+    let options = parser.Parse (inputs = argv, configurationReader = EnvironmentVariableConfigurationReader ())
+
+    let action = 
+      match (options.TryGetResult Down) with
+      | Some d -> FSharp.Data.Migrations.Down d
+      | None -> FSharp.Data.Migrations.Action.Up
+
+    // Return a record with values or default's set
+    {|
+      ConnectionString = options.GetResult Connection_String
+      MigrationsFolder = options.GetResult (Migrations_Folder, defaultValue = @"..\..\..\..\..\migrations")
+      Action = action
+    |}