+++ src/FSharp.Data.Migrations/PostgresProvider.fs
@@ -17,7 +17,7 @@ type PostgreSQL () =
       |> sb.add "CREATE TABLE "
       |> sb.addSomeF "%s." options.DbSchema
       |> sb.add options.DbTableName
-      |> sb.add "( script varchar(1024) NOT NULL PRIMARY KEY, created_at timestamp with time zone DEFAULT now() ) "
+      |> sb.add "( script varchar(1024) NOT NULL PRIMARY KEY, created_at timestamp with time zone DEFAULT now() ); "
       |> sb.toString
           
     member __.GetMigrations options =
@@ -25,13 +25,21 @@ type PostgreSQL () =
       |> sb.add "SELECT script FROM "
       |> sb.addSomeF "%s." options.DbSchema
       |> sb.add options.DbTableName
-      |> sb.add " ORDER BY script" 
+      |> sb.add " ORDER BY script; " 
       |> sb.toString
 
-    member __.RecordMigration options =
+    member __.AddUpMigration options =
       sb.create ()
-      |> sb.add"INSERT INTO "
+      |> sb.add "INSERT INTO "
       |> sb.addSomeF "%s." options.DbSchema
       |> sb.add options.DbTableName
-      |> sb.add " (script, created_at) VALUES(@ScriptName, now() );"
+      |> sb.add " (script, created_at) VALUES(@ScriptName, now() ); "
+      |> sb.toString
+      
+    member __.DeleteDownMigration options =
+      sb.create ()
+      |> sb.add "DELETE FROM "
+      |> sb.addSomeF "%s." options.DbSchema
+      |> sb.add options.DbTableName
+      |> sb.add " WHERE script = @ScriptName; "
       |> sb.toString